Test Login added Modified
Rename test class and update login URL and selector

Changed class name from `Tests` to `TestLogin` for clarity.
Updated the URL in the `GotoAsync` method to point to
"http://eaapp.somee.com/" and modified the click selector
from "text= Accept" to "text= Login" to align with the
new login process.